Refactor connection ID extraction modules
